requirements.txt
asgiref==3.9.1
Django==5.0.14
sqlparse==0.5.3
pandas==2.2.3
pyarrow==17.0.0
psycopg2-binary==2.9.10
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
drf-spectacular-sidecar==2025.8.1
whitenoise==6.7.0
google-cloud-storage==2.10.0
django-storages[google]==1.14.2
asgiref==3.9.1
attrs==25.3.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
Django==5.0.14
django-storages==1.14.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
drf-spectacular-sidecar==2025.8.1
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-core==2.4.3
google-cloud-storage==2.10.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
idna==3.10
inflection==0.5.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
numpy==2.3.2
pandas==2.2.3
proto-plus==1.26.1
protobuf==6.31.1
psycopg2-binary==2.9.10
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
rpds-py==0.27.0
rsa==4.9.1
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
whitenoise==6.7.0
celery==5.3.4
redis==5.0.1
